scripts/post-fs-data: allow disabling anti-bootloop
I have no idea what the use is but ok.

this is a callout to flahsrom's "force_I_want_a_brick" :)

for: https://github.com/backslashxx/mountify/issues/18

+++ module/post-fs-data.sh
@@ -52,7 +52,7 @@ BOOTCOUNT=0
 
 BOOTCOUNT=$(( BOOTCOUNT + 1))
 
-if [ $BOOTCOUNT -gt 1 ]; then
+if [ ! -f "$PERSISTENT_DIR/explicit_I_want_a_bootloop" ] && [ $BOOTCOUNT -gt 1 ]; then
 	touch $MODDIR/disable
 	rm "$MODDIR/count.sh"
 	string="description=anti-bootloop triggered. module disabled. enable to activate."
